DjangoApp - Copy/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-extensions==3.0.9
psycopg2==2.8.6
pydot==1.4.1
pyparsing==2.4.7
pytz==2020.1
sqlparse==0.3.1
uWSGI==2.0.19.1
whitenoise==5.2.0
django-crispy-forms==1.9.2
django-leaflet==0.27.1
django-pwa==1.0.10
djangorestframework==3.12.2
django-cors-headers==3.7.0
newsapi-python==0.2.6